Add pad bookmarks to start page
