+++ ep.json
@@ -5,8 +5,10 @@
       "hooks": {
         "eejsBlock_editbarMenuRight": "ep_bookmark/hooks:eejsBlock_editbarMenuRight",
         "eejsBlock_styles": "ep_bookmark/hooks:eejsBlock_styles",
+        "eejsBlock_scripts": "ep_bookmark/hooks:eejsBlock_scripts",
         "eejsBlock_body": "ep_bookmark/hooks:eejsBlock_body",
-        "eejsBlock_editorContainerBox": "ep_bookmark/hooks:eejsBlock_editorContainerBox"
+        "eejsBlock_editorContainerBox": "ep_bookmark/hooks:eejsBlock_editorContainerBox",
+        "eejsBlock_indexWrapper": "ep_bookmark/hooks:eejsBlock_indexWrapper"
       },
       "client_hooks": {
         "documentReady": "ep_bookmark/static/js/index",
